requirements.txt
bs4==0.0.2
cohere==5.5.8
faiss-cpu==1.8.0.post1
google-generativeai==0.7.0
huggingface-hub==0.23.4
langchain==0.2.6
langchain-cohere==0.1.8
langchain-community==0.2.6
langchain-core==0.2.10
langchain-experimental==0.0.62
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.1
langchainhub==0.1.20
pydantic==2.7.4
pypdf==4.2.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
streamlit==1.36.0
sentence-transformers
tabulate
